backend/requirements/base.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<4.0
django-filter>=24.0,<25.0
django-cors-headers>=4.4,<5.0
django-environ>=0.11,<1.0
django-postgresql-netfields>=1.3,<2.0
drf-nested-routers>=0.94,<1.0
psycopg[binary]>=3.2,<4.0
django-redis>=5.4,<6.0
gunicorn>=22.0,<23.0
whitenoise>=6.7,<7.0
celery>=5.4,<6.0
weasyprint>=62.0,<63.0
openpyxl>=3.1,<4.0
